feat: add `nargo init` command
